# backend/api-gateway-go/swagger/panda-gateway.yaml
openapi: 3.0.2
info:
  title: PANDA API Gateway
  description: This is an API Gateway to the PANDA database. This is the only place to access data from the PANDA database.
  version: "1.0"
  contact:
    name: Jiří Švácha
    email: jiri.svacha@eli-beams.eu
servers:
  - url: http://localhost:50000/v1
tags:
  - name: Security
    description: Everything about authentication and authorization
components:
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "You can get the JWT if you successfully authenticate via /authenticate POST endpoint"
  schemas:
    UserCredentials:
      type: object
      properties:
        username:
          type: string
          example: albert.einstein@eli-laser.eu
        password:
          type: string
          example: E=mc2
    UserInfo:
      type: object
      properties:
        uid:
          type: string
          example: 71864520-9e86-427c-901c-0c220f951775
        username:
          type: string
          example: albert.einstein@eli-laser.eu
        fullName:
          type: string
          example: Albert Einstein
        facility:
          type: string
          example: ELI ERIC
    ResponseMessage:
      type: object
      properties:
        message:
          type: string
          example: System was successfully created.
    AccessToken:
      type: object
      properties:
        access-token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyLCJleHAiOiIxMjM0NTY3ODk5In0.4UuocEvJTR3c1t2kB4f5lbRKfs5yKe7dW1Qtm_LITII
paths:
  /authenticate:
    post:
      summary: Authentication of the user
      description: Authenticate user by username and password and return JWT (json web token)
      operationId: AuthenticateByUsernameAndPassword
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCredentials"
      responses:
        "500":
          description: General server error
        "401":
          description: Unauthorized
        "200":
          description: Successfully authenticated. Returns JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessToken"
  /refresh-token:
    post:
      summary: Reauthentication of the user
      description: Reauthentication because of the token expiration. Authenticate user by JWT (json web token) presented in the Authorization header and if the JWT is still ok create and returns new JWT.
      operationId: RefreshToken
      security:
        - jwtAuth: []
      tags:
        - Security
      responses:
        "500":
          description: General server error
        "401":
          description: Unauthorized
        "200":
          description: Successfully reauthenticated. Returns new JWT.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessToken"
  /authenticated-user:
    get:
      summary: Get user info by JWT presented in header
      description: Get user info by JWT (json we token) presented in Authorization header in the request
      operationId: GetUserByJWT
      security:
        - jwtAuth: []
      tags:
        - Security
      responses:
        "500":
          description: General server error
        "401":
          description: Unauthorized
        "200":
          description: Returns user information from the JWT that was successfully chekced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"